Added min-width to images in collection
Gave images min-width so that the width of each item before the image has lazy loading in is not zero. Items having a width of zero caused all items to be visible in the viewport when the collection was loaded so they all get lazy loaded in when you open the image collections.

Also removed modalButton class from close button in the image collection modals as I didn't want to give the close button a background. The width of the options buttons has been changed to  stop the options buttons in the page from overlapping the close button